+++ km3modules/plot.py
@@ -15,9 +15,6 @@ import time
 
 import pandas as pd
 import numpy as np
-import matplotlib
-# Force matplotlib to not use any Xwindows backend.
-matplotlib.use('Agg')
 import matplotlib.pyplot as plt    # noqa
 import matplotlib.ticker as ticker
 from matplotlib import pylab    # noqa
@@ -334,7 +331,7 @@ def ztplot(
                                 for m in markers[1:]] + ['triggered'],
             scatterpoints=1,
             markerscale=1,
-            loc='upper left',
+            loc='lower right',
             frameon=True,
             framealpha=0.7
         )
@@ -366,3 +363,44 @@ def trim_axes(axes, n):
     for ax in axes[n:]:
         ax.remove()
     return axes[:n]
+
+
+def cumulative_run_livetime(qtable, kind='runs'):
+    """Create a figure which plots the cumulative livetime of runs
+
+    Parameters
+    ----------
+    qtable: pandas.DataFrame
+        A table which has the run number as index and columns for
+        'livetime_s', 'timestamp' and 'datetime' (pandas datetime).
+    kind: str
+        'runs' to plot for each run or 'timeline' to plot based
+        on the actual run time.
+
+    Returns
+    -------
+    matplotlib.Figure
+    """
+    fig, ax = plt.subplots()
+
+    options = {
+        'runs': {
+            'xlabel': 'run',
+            'xs': qtable.index
+        },
+        'timeline': {
+            'xlabel': None,
+            'xs': qtable.datetime
+        }
+    }
+
+    actual_livetime = np.cumsum(qtable['livetime_s'])
+    optimal_livetime = np.cumsum(qtable.timestamp.diff())
+
+    ax.plot(options[kind]['xs'], actual_livetime, label='actual livetime')
+    ax.plot(options[kind]['xs'], optimal_livetime, label='100% livetime')
+    ax.set_xlabel(options[kind]['xlabel'])
+    ax.set_ylabel('time / s')
+    ax.legend()
+
+    return fig